Use the new feature flag control for the 'multi-agent' training settings
resolves #1265
