+++ pathmind-services/src/test/java/io/skymind/pathmind/services/project/rest/ModelAnalyzerApiClientTest.java
@@ -2,7 +2,10 @@ package io.skymind.pathmind.services.project.rest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.skymind.pathmind.services.project.rest.dto.HyperparametersDTO;
+import io.skymind.pathmind.shared.featureflag.Feature;
+import io.skymind.pathmind.shared.featureflag.FeatureManager;
 import io.skymind.pathmind.shared.utils.ObjectMapperHolder;
+import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.core.io.ClassPathResource;
@@ -12,9 +15,18 @@ import java.io.File;
 import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.*;
 
 @Ignore
 public class ModelAnalyzerApiClientTest {
+
+    FeatureManager featureManager = mock(FeatureManager.class);
+
+    @Before
+    public void setUp() {
+        reset(featureManager);
+        when(featureManager.isEnabled(eq(Feature.MULTI_AGENT_TRAINING))).thenReturn(false);
+    }
 
     @Test
     public void testBasicModelAnalyze() throws IOException {
@@ -22,7 +34,7 @@ public class ModelAnalyzerApiClientTest {
 
         final ModelAnalyzerApiClient client = new ModelAnalyzerApiClient(
                 "https://ma.dev.devpathmind.com",
-                null,false,
+                null,featureManager,
                 objectMapper,
                 WebClient.builder());
