+++ pathmind-shared/src/main/java/io/skymind/pathmind/shared/featureflag/Feature.java
@@ -1,0 +1,6 @@
+package io.skymind.pathmind.shared.featureflag;
+
+public enum Feature {
+    REWARD_VARIABLES_FEATURE,
+    MULTI_AGENT_TRAINING
+}
