--- pathmind-webapp/src/main/java/io/skymind/pathmind/webapp/security/Feature.java
@@ -1,5 +1,0 @@
-package io.skymind.pathmind.webapp.security;
-
-public enum Feature {
-    REWARD_VARIABLES_FEATURE
-}
