+++ pathmind-webapp/src/main/java/io/skymind/pathmind/webapp/ui/views/experiment/ExperimentView.java
@@ -41,8 +41,8 @@ import io.skymind.pathmind.shared.security.Routes;
 import io.skymind.pathmind.shared.utils.PolicyUtils;
 import io.skymind.pathmind.webapp.data.utils.ExperimentUtils;
 import io.skymind.pathmind.webapp.exception.InvalidDataException;
-import io.skymind.pathmind.webapp.security.Feature;
-import io.skymind.pathmind.webapp.security.FeatureManager;
+import io.skymind.pathmind.shared.featureflag.Feature;
+import io.skymind.pathmind.shared.featureflag.FeatureManager;
 import io.skymind.pathmind.webapp.ui.components.TabPanel;
 import io.skymind.pathmind.webapp.ui.components.navigation.Breadcrumbs;
 import io.skymind.pathmind.webapp.ui.components.notesField.NotesField;
